backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pymongo>=4.10
motor>=3.6
pydantic>=2.9
python-dotenv>=1.0
websockets>=13.1
httpx>=0.27
cloudinary>=1.40
opencv-python-headless>=4.10
numpy>=1.26
